requirements/base.txt
django==4.0
python-decouple==3.8
djangorestframework==3.14.0
django-fsm==2.8.1
requests==2.28.2
django-redis==5.2.0
autopep8==2.0.2
